+++ src/dataloader.hpp
@@ -45,7 +45,7 @@ class DataLoader {
         virtual ~DataLoader() {}
 
         void init(string rootpth, string fname, vector<int> sz, bool is_train);
-        void _get_batch(vector<float>* &data, vector<int>& size, vector<int>* &labels);
+        void _get_batch(vector<float>* &data, vector<int>& size, vector<int64_t>* &labels);
         void _shuffle();
         void _start();
         void _restart();
